requirements.txt
streamlit
streamlit-drawable-canvas
opencv-python
numpy
keras
sympy
google-generativeai
streamlit-lottie
